New approach to registry
Using the constant namespace is neat but it's not practical. At least,
not the way I have it implementd at the moment. This new approach uses a
singleton class `BM_Registry` and - importantly - forces all accesses to
be specifically _by_name or _by_obj.

Twice now I've tried to make a pithy interface that allowed overloading
of a single "registered?" function, but it ended up a mess in both
cases. This approach is already feeling much cleaner as evidenced by the
tidy test case and class definition.